requirements.txt
smartapi-python
firebase-admin
pytz
logzero
websocket-client
pyotp
